Break controlpanel so that Plone 5 at least starts up.
The plone.app.controlpanel stuff that we use as base has been completely removed and reworked,
so we need extensive changes.
We would probably want to rework our stuff to be based on plone.app.registry.
A site migrated from Plone 4.3 *does* work, but the controlpanel is broken.
A new Plone 5 site probably works as well, but again, the controlpanel is broken.